Refactor NextAuth configuration to eliminate code duplication
- Created shared auth-config.ts with createAuthConfig() function
- Refactored lib/auth.ts from 102 lines to 4 lines using shared config
- Refactored api/auth/[...nextauth]/route.ts from 117 lines to 6 lines
- Fixed middleware.ts by adding missing PUBLIC_ROUTES and PROTECTED_ROUTES constants
- Improved DRY principles by eliminating duplicated NextAuth provider logic
- All builds now pass successfully
