+++ apps/web/app/api/auth/[...nextauth]/route.ts
@@ -1,116 +1,6 @@
 import NextAuth from "next-auth";
-import Google from "next-auth/providers/google";
-import GitHub from "next-auth/providers/github";
-import Credentials from "next-auth/providers/credentials";
-import { PrismaAdapter } from "@auth/prisma-adapter";
-import { prisma } from "../../../../lib/prisma";
-import bcrypt from "bcryptjs";
-import { signInSchema } from "../../../../lib/zod";
+import { createAuthConfig } from "../../../../lib/auth-config";
 
-const providers = [];
-
-// Only add Google provider if credentials are available
-if (process.env.AUTH_GOOGLE_ID && process.env.AUTH_GOOGLE_SECRET) {
-  providers.push(
-    Google({
-      clientId: process.env.AUTH_GOOGLE_ID,
-      clientSecret: process.env.AUTH_GOOGLE_SECRET,
-      authorization: { params: { access_type: "offline", prompt: "consent" } },
-    })
-  );
-}
-
-// Only add GitHub provider if credentials are available
-if (process.env.AUTH_GITHUB_ID && process.env.AUTH_GITHUB_SECRET) {
-  providers.push(
-    GitHub({
-      clientId: process.env.AUTH_GITHUB_ID,
-      clientSecret: process.env.AUTH_GITHUB_SECRET,
-    })
-  );
-}
-
-// Always add credentials provider for email/password auth
-providers.push(
-  Credentials({
-    credentials: {
-      email: { label: "Email", type: "email" },
-      password: { label: "Password", type: "password" },
-    },
-    authorize: async (credentials) => {
-      try {
-        const { email, password } = await signInSchema.parseAsync(credentials);
-
-        // Find user in database
-        const user = await prisma.user.findUnique({
-          where: { email },
-        });
-
-        if (!user || !user.password) {
-          throw new Error("Invalid credentials.");
-        }
-
-        // Verify password
-        const isValidPassword = await bcrypt.compare(password, user.password);
-
-        if (!isValidPassword) {
-          throw new Error("Invalid credentials.");
-        }
-
-        return {
-          id: user.id,
-          email: user.email,
-          name: user.name,
-          image: user.image,
-        };
-      } catch {
-        return null;
-      }
-    },
-  })
-);
-
-const handler = NextAuth({
-  adapter: PrismaAdapter(prisma),
-  session: {
-    strategy: "jwt",
-  },
-  providers,
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.id = user.id;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (token && session.user) {
-        session.user.id = token.id as string;
-      }
-      return session;
-    },
-    async signIn({ account, profile }) {
-      if (account?.provider === "google" && profile) {
-        return (profile as { email_verified?: boolean })?.email_verified === true;
-      }
-      if (account?.provider === "github") {
-        return true;
-      }
-      return true;
-    },
-  },
-  pages: {
-    signIn: "/signin",
-  },
-  secret:
-    process.env.AUTH_SECRET ||
-    (process.env.NODE_ENV === "production"
-      ? (() => {
-          throw new Error(
-            "AUTH_SECRET environment variable must be set in production."
-          );
-        })()
-      : "development-secret"),
-});
+const handler = NextAuth(createAuthConfig());
 
 export { handler as GET, handler as POST };
