+++ apps/web/lib/auth.ts
@@ -1,101 +1,4 @@
 import NextAuth from "next-auth";
-import Google from "next-auth/providers/google";
-import GitHub from "next-auth/providers/github";
-import Credentials from "next-auth/providers/credentials";
-import { PrismaAdapter } from "@auth/prisma-adapter";
-import { prisma } from "./prisma";
-import bcrypt from "bcryptjs";
-import { signInSchema } from "./zod";
+import { createAuthConfig } from "./auth-config";
 
-export const { handlers, auth, signIn, signOut } = NextAuth({
-  adapter: PrismaAdapter(prisma),
-  session: {
-    strategy: "jwt",
-  },
-  providers: [
-    ...(process.env.AUTH_GOOGLE_ID && process.env.AUTH_GOOGLE_SECRET 
-      ? [Google({
-          clientId: process.env.AUTH_GOOGLE_ID,
-          clientSecret: process.env.AUTH_GOOGLE_SECRET,
-          authorization: { params: { access_type: "offline", prompt: "consent" } },
-        })]
-      : []
-    ),
-    ...(process.env.AUTH_GITHUB_ID && process.env.AUTH_GITHUB_SECRET
-      ? [GitHub({
-          clientId: process.env.AUTH_GITHUB_ID,
-          clientSecret: process.env.AUTH_GITHUB_SECRET,
-        })]
-      : []
-    ),
-    Credentials({
-      credentials: {
-        email: { label: "Email", type: "email" },
-        password: { label: "Password", type: "password" },
-      },
-      authorize: async (credentials) => {
-        try {
-          const { email, password } =
-            await signInSchema.parseAsync(credentials);
-
-          // Find user in database
-          const user = await prisma.user.findUnique({
-            where: { email },
-          });
-
-          if (!user || !user.password) {
-            throw new Error("Invalid credentials.");
-          }
-
-          // Verify password
-          const isValidPassword = await bcrypt.compare(password, user.password);
-
-          if (!isValidPassword) {
-            throw new Error("Invalid credentials.");
-          }
-
-          return {
-            id: user.id,
-            email: user.email,
-            name: user.name,
-            image: user.image,
-          };
-        } catch {
-          return null;
-        }
-      },
-    }),
-  ],
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.id = user.id;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (token && session.user) {
-        session.user.id = token.id as string;
-      }
-      return session;
-    },
-    async signIn({ account, profile }) {
-      if (account?.provider === "google" && profile) {
-        return (profile as { email_verified?: boolean })?.email_verified === true;
-      }
-      if (account?.provider === "github") {
-        return true;
-      }
-      return true;
-    },
-  },
-  pages: {
-    signIn: "/signin",
-  },
-  secret: (() => {
-    if (process.env.NODE_ENV === "production" && !process.env.AUTH_SECRET) {
-      throw new Error("AUTH_SECRET must be set in production environments.");
-    }
-    return process.env.AUTH_SECRET || "development-secret";
-  })(),
-});
+export const { handlers, auth, signIn, signOut } = NextAuth(createAuthConfig());
